// src/types/comprobante.ts
export default interface Comprobante {
    codcomp: number;
    tipoc: string;
    nombre: string;
    apellidos:string;
    tipodoc: string;
    nrodoc: string;
    ruc: string;
    razons: string;
    paisr: string;
    dircalle: string;
    apartamento: string;
    distrito: string;
    telefono: string;
    correo: string;
    fechaemi: string;
    subtotal: number;
    igv: number;
    total: number;
    notas: string;



    
}